requirements.txt
librosa==0.9.1
# numpy
opencv-contrib-python
opencv-python
# torchvision
tqdm
numba
moviepy
minio
pydub
fastapi==0.68.0
uvicorn==0.15.0
pydantic==1.10.18
python-dotenv
Cython==0.29.21
# librosa==0.8.0
matplotlib==3.3.1
numpy==1.18.5
phonemizer==2.2.1
scipy==1.5.2
tensorboard==2.3.0
torch==1.6.0
torchvision==0.7.0
Unidecode==1.1.1